added triangles, worked on updating renderer capsulation

+++ src/objects/march.rs
@@ -6,7 +6,7 @@ use crate::structures::vec3::Vec3;
 use crate::structures::ray::Ray;
 
 use crate::render::EPSILON;
-pub const MAX_STEPS: u32 = 64;
+pub const MAX_STEPS: usize = 64;
 pub const MAX_DEPTH: f64 = 40.0;
 
 pub trait March {
